Schema-checker and simple evaluator
Introduced a schema-checker, to allow the front-end driver to enforce
the interfaces between phases. This had a large knock-on effect
throughout the code, as the first phase, the macro-expander, had to be
augmented to become a parser as well (which is actually how it should
have been done in the first place), the parser has been changed to a
simple internal-definition-rewriting phase, and many other phases were
simplified or changed shape to be more easily understandable.

The (node-tree-map!) routine has been changed to be
(node-children-map!), which allows for more flexible traversal of node
trees. Consequently, the annotation phase can now make use of this
generic tree-walking utility.

Arginfo structures and capture records have been formalised, and use
the same node primitive as is used to represent other kinds of tree
and graph, so that they can be schema-checked as well. This cleaned up
the annotation phase no end.

I added a simple evaluator for the tree structure emitted by the
complete front-end. It serves two purposes: to act as an evaluator for
the macro-expander (perhaps just temporarily, but perhaps not; I am
still undecided), and to clarify a few implementation issues regarding
writing backends. The code makes a good starting point for a new
backend.

I haven't yet incorporated Kelsey's distinctions between lambda-cont,
lambda-proc and lambda-jump, which he uses to great effect in his 1995
paper on the correspondence between CPS and SSA, but I plan to once I
get the compiler bootstrapped. Something of the kind will be required
to enable some aggressive optimisations.

The case-lambda implementation has been removed. It can be reinserted
later. Perhaps a partial-evaluator will help make it unnecessary?

*Lots* of the new code is untested. A unit- and regression-test suite
 is becoming urgent.

darcs-hash:20050123151852-51b3b-5603a40e11023ddc67f675e4dd8add702c3ac484.gz